requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
deprecation==2.1.0
distro==1.8.0
exceptiongroup==1.2.0
fonttools==4.45.0
gotrue==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-resources==6.1.1
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.7.4
numpy==1.24.4
openai==1.3.5
packaging==23.2
Pillow==10.1.0
postgrest==0.13.0
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
realtime==1.0.0
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
storage3==0.7.0
StrEnum==0.4.15
supabase==2.1.0
supafunc==0.3.1
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
websockets==10.4
zipp==3.17.0
